+++ profiles/local-frontend.js
@@ -3,13 +3,15 @@ const APP_ID = 'ocp-advisor';
 const FRONTEND_PORT = 8002;
 const routes = {};
 
-routes[`/beta/${SECTION}/${APP_ID}`] = {
+routes[`/preview/${SECTION}/${APP_ID}`] = {
   host: `https://localhost:${FRONTEND_PORT}`,
 };
 routes[`/${SECTION}/${APP_ID}`] = {
   host: `https://localhost:${FRONTEND_PORT}`,
 };
-routes[`/beta/apps/${APP_ID}`] = { host: `https://localhost:${FRONTEND_PORT}` };
+routes[`/preview/apps/${APP_ID}`] = {
+  host: `https://localhost:${FRONTEND_PORT}`,
+};
 routes[`/apps/${APP_ID}`] = { host: `https://localhost:${FRONTEND_PORT}` };
 
 module.exports = { routes };