+++ src/Components/Cluster/Cluster.cy.js
@@ -1,214 +1,186 @@
 import React from 'react';
-import { mount } from '@cypress/react';
-import { MemoryRouter } from 'react-router-dom';
-
-import { Intl } from '../../Utilities/intlHelper';
-import { Cluster } from './Cluster';
-import { Provider } from 'react-redux';
-import getStore from '../../Store';
+import Cluster from '.';
 import singleClusterPageReport from '../../../cypress/fixtures/api/insights-results-aggregator/v2/cluster/dcb95bbf-8673-4f3a-a63c-12d4a530aa6f/reports-disabled-false.json';
 import {
   checkNoMatchingRecs,
   checkRowCounts,
 } from '../../../cypress/utils/table';
+import {
+  clusterReportsInterceptors as interceptors,
+  upgradeRisksInterceptors,
+} from '../../../cypress/utils/interceptors';
 
 // selectors
 const CLUSTER_HEADER = '#cluster-header';
 const BREADCRUMBS = 'nav[class=pf-c-breadcrumb]';
 const RULES_TABLE = '#cluster-recs-list-table';
 const FILTER_CHIPS = 'li[class=pf-c-chip-group__list-item]';
-let props;
+const ALERT = '[data-ouia-component-type="PF4/Alert"]';
+
+const CLUSTER_ID = '123';
+const CLUSTER_NAME = 'Cluster With Issues';
+
+const mount = (initialEntries = ['/clusters/123']) => {
+  cy.mountWithContext(<Cluster />, {
+    path: '/clusters/:clusterId',
+    routerProps: { initialEntries },
+  });
+};
 
 describe('cluster page', () => {
-  beforeEach(() => {
-    // the flag tells not to fetch external federated modules
-    window.CYPRESS_RUN = true;
+  describe('in the successful state', () => {
+    beforeEach(() => {
+      interceptors.successful();
+    });
 
-    props = {
-      cluster: {
-        isError: false,
-        isUninitialized: false,
-        isLoading: false,
-        isFetching: false,
-        isSuccess: true,
-        data: singleClusterPageReport,
-      },
-      displayName: {
-        data: singleClusterPageReport.report.meta.cluster_name,
-      },
-      clusterId: 'foobar',
-    };
+    it('renders main components', () => {
+      mount();
+
+      // renders breadcrumbs
+      cy.get(BREADCRUMBS)
+        .should('have.length', 1)
+        .get('.pf-c-breadcrumb__list > :nth-child(2)')
+        .should('have.text', CLUSTER_NAME);
+      // renders cluster header
+      cy.get(CLUSTER_HEADER).should('have.length', 1);
+      // renders table component
+      cy.get(RULES_TABLE).should('have.length', 1);
+      // test how many rows were rendered
+      checkRowCounts(singleClusterPageReport.report.data.length, true);
+    });
+
+    it('adds additional filters passed by the query parameters, 1', () => {
+      mount(['/clusters/123?total_risk=1&text=foo+bar&category=2']);
+
+      cy.get(BREADCRUMBS).should('have.length', 1);
+      cy.get(CLUSTER_HEADER).should('have.length', 1);
+      cy.get(RULES_TABLE);
+      cy.get(FILTER_CHIPS).each(($el) =>
+        expect($el.text()).to.be.oneOf(['foo bar', 'Low', 'Performance'])
+      );
+      checkNoMatchingRecs();
+    });
+
+    it('adds additional filters passed by the query parameters, 2', () => {
+      mount(['/clusters/123?total_risk=2&text=foo&category=1']);
+
+      cy.get(BREADCRUMBS).should('have.length', 1);
+      cy.get(CLUSTER_HEADER).should('have.length', 1);
+      cy.get(RULES_TABLE);
+      cy.get(FILTER_CHIPS).each(($el) =>
+        expect($el.text()).to.be.oneOf([
+          'foo',
+          'Moderate',
+          'Service Availability',
+        ])
+      );
+      checkNoMatchingRecs();
+    });
+
+    it('last breadcrumb is cluster name', () => {
+      mount();
+
+      cy.get(BREADCRUMBS)
+        .should('have.length', 1)
+        .get('.pf-c-breadcrumb__list > :nth-child(2)')
+        .should('have.text', CLUSTER_NAME);
+    });
+
+    it('last breadcrumb is cluster id when name is not available', () => {
+      interceptors['successful, cluster name is null']();
+      mount();
+
+      cy.get(BREADCRUMBS)
+        .should('have.length', 1)
+        .get('.pf-c-breadcrumb__list > :nth-child(2)')
+        .should('have.text', CLUSTER_ID);
+    });
   });
 
-  it('cluster page in the successful state', () => {
-    mount(
-      <MemoryRouter>
-        <Intl>
-          <Provider store={getStore()}>
-            <Cluster {...props} />
-          </Provider>
-        </Intl>
-      </MemoryRouter>
-    );
-    // renders breadcrumbs
-    cy.get(BREADCRUMBS)
-      .should('have.length', 1)
-      .get('.pf-c-breadcrumb__list > :nth-child(2)')
-      .should('have.text', 'Cluster With Issues');
-    // renders cluster header
-    cy.get(CLUSTER_HEADER).should('have.length', 1);
-    // renders table component
-    cy.get(RULES_TABLE).should('have.length', 1);
-    // test how many rows were rendered
-    checkRowCounts(singleClusterPageReport.report.data.length, true);
+  describe('in the loading state', () => {
+    beforeEach(() => {
+      interceptors['long responding']();
+    });
+
+    it('renders skeleton', () => {
+      mount();
+
+      // renders breadcrumbs
+      cy.get(BREADCRUMBS).should('have.length', 1);
+      // renders cluster header
+      cy.get(CLUSTER_HEADER).should('have.length', 1);
+      // renders table component
+      cy.get(RULES_TABLE).should('have.length', 1);
+    });
   });
 
-  it('cluster page in the loading state', () => {
-    props = {
-      ...props,
-      cluster: {
-        ...props.cluster,
-        isFetching: true,
-        isSuccess: false,
-        data: undefined,
-      },
-    };
-    mount(
-      <MemoryRouter>
-        <Intl>
-          <Provider store={getStore()}>
-            <Cluster {...props} />
-          </Provider>
-        </Intl>
-      </MemoryRouter>
-    );
-    // renders breadcrumbs
-    cy.get(BREADCRUMBS).should('have.length', 1);
-    // renders cluster header
-    cy.get(CLUSTER_HEADER).should('have.length', 1);
-    // renders table component
-    cy.get(RULES_TABLE).should('have.length', 1);
-    cy.ouiaId('loading-skeleton').should('have.length', 1);
-  });
+  describe('in the error state', () => {
+    beforeEach(() => {
+      interceptors['server error']();
+    });
 
-  it('cluster page in the error state', () => {
-    props = {
-      ...props,
-      cluster: {
-        ...props.cluster,
-        isError: true,
-        isSuccess: false,
-        isFetching: false,
-        data: undefined,
-      },
-    };
-    mount(
-      <MemoryRouter>
-        <Intl>
-          <Provider store={getStore()}>
-            <Cluster {...props} />
-          </Provider>
-        </Intl>
-      </MemoryRouter>
-    );
-    // renders breadcrumbs
-    cy.get(BREADCRUMBS).should('have.length', 1);
-    // renders cluster header
-    cy.get(CLUSTER_HEADER).should('have.length', 1);
-    // renders table component
-    cy.get(RULES_TABLE).should('have.length', 1);
-    cy.get('.pf-c-empty-state').should('have.length', 1);
-  });
+    it('renders empty state component', () => {
+      mount();
 
-  it('adds additional filters passed by the query parameters №1', () => {
-    mount(
-      <MemoryRouter initialEntries={['?total_risk=1&text=foo+bar&category=2']}>
-        <Intl>
-          <Provider store={getStore()}>
-            <Cluster {...props} />
-          </Provider>
-        </Intl>
-      </MemoryRouter>
-    );
-    cy.get(BREADCRUMBS);
-    cy.get(CLUSTER_HEADER);
-    cy.get(RULES_TABLE);
-    cy.get(FILTER_CHIPS).each(($el) =>
-      expect($el.text()).to.be.oneOf(['foo bar', 'Low', 'Performance'])
-    );
-    checkNoMatchingRecs();
-  });
-
-  it('adds additional filters passed by the query parameters №2', () => {
-    mount(
-      <MemoryRouter initialEntries={['?total_risk=2&text=foo&category=1']}>
-        <Intl>
-          <Provider store={getStore()}>
-            <Cluster {...props} />
-          </Provider>
-        </Intl>
-      </MemoryRouter>
-    );
-    cy.get(BREADCRUMBS);
-    cy.get(CLUSTER_HEADER);
-    cy.get(RULES_TABLE);
-    cy.get(FILTER_CHIPS).each(($el) =>
-      expect($el.text()).to.be.oneOf([
-        'foo',
-        'Moderate',
-        'Service Availability',
-      ])
-    );
-    checkNoMatchingRecs();
+      // renders breadcrumbs
+      cy.get(BREADCRUMBS).should('have.length', 1);
+      // renders cluster header
+      cy.get(CLUSTER_HEADER).should('have.length', 1);
+      // renders table component
+      cy.get(RULES_TABLE).should('have.length', 1);
+      cy.get('.pf-c-empty-state').should('have.length', 1);
+    });
   });
 });
 
-describe('Cluster page display name test №1', () => {
-  before(() => {
-    props = {
-      cluster: {
-        isError: false,
-        isUninitialized: false,
-        isLoading: false,
-        isFetching: false,
-        isSuccess: true,
-        data: singleClusterPageReport,
-      },
-      clusterId: 'Cluster Id',
-    };
+describe('upgrade risks banner', () => {
+  it('has some upgrade risks', () => {
+    upgradeRisksInterceptors.successful();
+    mount();
+
+    cy.get(ALERT).should('have.class', 'pf-m-warning');
+    cy.get(ALERT).within(() => {
+      cy.get('h4').should('contain.text', 'Resolve upgrade risks');
+    });
   });
 
-  it('Cluster breadcrumbs name should be Cluster With Issues', () => {
-    mount(
-      <MemoryRouter>
-        <Intl>
-          <Provider store={getStore()}>
-            <Cluster {...props} />
-          </Provider>
-        </Intl>
-      </MemoryRouter>
-    );
-    cy.get(BREADCRUMBS)
-      .should('have.length', 1)
-      .get('.pf-c-breadcrumb__list > :nth-child(2)')
-      .should('have.text', 'Cluster With Issues');
+  it('has no upgrade risks', () => {
+    upgradeRisksInterceptors['successful, empty']();
+    mount();
+
+    cy.get(ALERT).should('have.class', 'pf-m-success');
+    cy.get(ALERT).within(() => {
+      cy.get('h4').should(
+        'contain.text',
+        'No known upgrade risks identified for this cluster.'
+      );
+    });
   });
 
-  it('Cluster breadcrumbs name should be = Cluster Id', () => {
-    mount(
-      <MemoryRouter>
-        <Intl>
-          <Provider store={getStore()}>
-            <Cluster
-              {...{ ...props, cluster: { ...props.cluster, data: null } }}
-            />
-          </Provider>
-        </Intl>
-      </MemoryRouter>
-    );
-    cy.get(BREADCRUMBS)
-      .should('have.length', 1)
-      .get('.pf-c-breadcrumb__list > :nth-child(2)')
-      .should('have.text', 'Cluster Id');
+  it('upgrade risks not found', () => {
+    upgradeRisksInterceptors['error, not found']();
+    mount();
+
+    cy.get(ALERT).should('have.class', 'pf-m-warning');
+    cy.get(ALERT).within(() => {
+      cy.get('h4').should(
+        'contain.text',
+        'Upgrade risks are not currently available.'
+      );
+    });
+  });
+
+  it('should not render alert in other error', () => {
+    upgradeRisksInterceptors['error, other']();
+    mount();
+
+    cy.get(ALERT).should('not.exist');
+  });
+
+  it('should not render alert in the loading state', () => {
+    upgradeRisksInterceptors['long responding']();
+    mount();
+
+    cy.get(ALERT).should('not.exist');
   });
 });