+++ src/Components/UpgradeRisksTable/index.js
@@ -1,0 +1,1 @@
+export { default as UpgradeRisksTable } from './UpgradeRisksTable';